unittests for turning learning on and off between simulate calls and setup calls
